Updated text area component
As of right now when isError is present but no error message is present, we do not display the error message span.

This has been updated to always display the span. The idea is that we will now use the before element in this span
to create the error icon. This guarantees that the icon will be between the textarea and error message, making it
easier to position it in a reliable way while letting the error message increase the size of the container to push
the content below it.